Apply rigid body physics on a dish.
Use _ready() for a purpose of initialization. A logic that changes position of rigid body should be in _integrate_forces rather than _process, because _process cannot affect rigid body physics.

Tried to show child hierarchy on Scene dock of editor. However, I failed. A print_tree_pretty() function shows custom classes hierarchy only in an output dock. Not in a scene dock. 

A forum article for showing hierarchy == https://forum.godotengine.org/t/when-creating-nodes-in-a-c-module-how-do-you-make-them-appear-as-a-hierarchy-in-the-editor/25674