#1 Implementing a refrigerator platform.
main.tscn - Implement it in Godot engine. In order to setup UI in main scene, added Canvas layer. It contains a NinePatchRect that a control displays a texture by keeping its corners intact. Then, it has a margin container to keep margins between textures and edges. Inside it, scroll container is  included in the margin container. It provided a scroll functionality. Finally Grid container has a lot of items to display items.

Helpful videos:
https://www.youtube.com/watch?v=8cV-5ByZLOE (Old drag&drop)
https://www.youtube.com/watch?v=lrAwX2t1mGY (Inventory visual setups)
https://www.youtube.com/watch?v=dZYlwmBCziM (Inventory tech setups)