#1 Implemented dragdrop trio functions. Figured how to spawn an object.
InventoryBg - PNG file shows background of inventory slot.
ItemBg.gd - A gdscript file that implemented dragdrop trio (_get_drag_data, _can_drop_data, and _drop_data). It worked if and only if target data has the trio function. Thus, it is impossible to spawn an object using it. I need to implement a spawner to spawn an object in game world.